Refactor metrics logic to metrics class
